+++ ai_launchpad/cleanup_template.sh
