+++ pyproject.toml
@@ -8,7 +8,7 @@ dependencies = ["requests>=2.31.0"]
 dev = ["pytest>=8.0", "pytest-cov>=6.0"]
 
 [tool.pytest.ini_options]
-addopts = "-v --cov=. --cov-report term-missing --cov-fail-under 95 --cov-config=pyproject.toml"
+addopts = "-v --cov=ai_launchpad --cov-report term-missing --cov-fail-under 95 --cov-config=pyproject.toml"
 
 python_files = "tests/test_*.py"
