+++ tests/test_agents.py
@@ -1,7 +1,7 @@
 import pytest
 from unittest.mock import patch
 
-from agents import BaseAgent, ClaudeAgent, CodexAgent, AGENT_REGISTRY, get_agent
+from ai_launchpad.agents import BaseAgent, ClaudeAgent, CodexAgent, AGENT_REGISTRY, get_agent
 
 
 class TestBaseAgent:
@@ -14,14 +14,14 @@ class TestBaseAgent:
         agent = BaseAgent()
         assert agent.cmd == ""
 
-    @patch("agents.read_config", return_value={})
+    @patch("ai_launchpad.agents.read_config", return_value={})
     def test_generate_prompt_tells_agent_to_read_task_md(self, _mock_config):
         agent = BaseAgent()
         prompt = agent.generate_prompt()
         assert "task.md" in prompt
 
     @patch(
-        "agents.read_config",
+        "ai_launchpad.agents.read_config",
         return_value={"custom_agent_message": "Be thorough!"},
     )
     def test_generate_prompt_with_custom_message(self, _mock_config):
@@ -29,7 +29,7 @@ class TestBaseAgent:
         prompt = agent.generate_prompt()
         assert "Be thorough!" in prompt
 
-    @patch("agents.read_config", return_value={})
+    @patch("ai_launchpad.agents.read_config", return_value={})
     def test_generate_prompt_does_not_contain_work_item_fields(self, _mock_config):
         agent = BaseAgent()
         prompt = agent.generate_prompt().lower()
@@ -38,7 +38,7 @@ class TestBaseAgent:
         assert "link:" not in prompt
 
     @patch(
-        "agents.read_config",
+        "ai_launchpad.agents.read_config",
         return_value={"custom_agent_message": "Be thorough!"},
     )
     def test_generate_prompt_with_custom_message_replaces_default(self, _mock_config):
@@ -46,7 +46,7 @@ class TestBaseAgent:
         prompt = agent.generate_prompt()
         assert "task.md" not in prompt
 
-    @patch("agents.read_config", return_value={})
+    @patch("ai_launchpad.agents.read_config", return_value={})
     def test_generate_prompt_without_custom_message__no_extra_blank_lines(
         self, _mock_config
     ):
