+++ tests/test_config.py
@@ -2,7 +2,7 @@ import pytest
 from pathlib import Path
 from unittest.mock import patch
 
-from config import _read_toml, _missing_required_fields, read_config
+from ai_launchpad.config import _read_toml, _missing_required_fields, read_config
 
 
 class TestReadToml:
@@ -48,9 +48,7 @@ class TestReadConfig:
         config_file.write_text(
             'base_worktrees_dir = "/tmp/ctx"\nbase_source_dir = "/tmp/src"\n'
         )
-        with patch("config.Path") as mock_path_cls:
-            # Path(__file__) is called, then .with_name("config.toml")
-            mock_path_cls.return_value.with_name.return_value = config_file
+        with patch("ai_launchpad.config._default_config_path", return_value=config_file):
             result = read_config()
         assert result["base_worktrees_dir"] == "/tmp/ctx"
         assert result["base_source_dir"] == "/tmp/src"
@@ -58,15 +56,13 @@ class TestReadConfig:
     def test_empty_config_raises(self, tmp_path):
         config_file = tmp_path / "config.toml"
         config_file.write_text("")
-        with patch("config.Path") as mock_path_cls:
-            mock_path_cls.return_value.with_name.return_value = config_file
+        with patch("ai_launchpad.config._default_config_path", return_value=config_file):
             with pytest.raises(ValueError, match="empty or invalid"):
                 read_config()
 
     def test_missing_required_fields_raises(self, tmp_path):
         config_file = tmp_path / "config.toml"
         config_file.write_text('base_worktrees_dir = "/tmp/ctx"\n')
-        with patch("config.Path") as mock_path_cls:
-            mock_path_cls.return_value.with_name.return_value = config_file
+        with patch("ai_launchpad.config._default_config_path", return_value=config_file):
             with pytest.raises(ValueError, match="base_source_dir"):
                 read_config()
